Se termina de implementar la funcionalidad del Token para acceder a datos.
