+++ backend/src/routes/pokemon.js
@@ -1,7 +1,8 @@
 const express = require("express");
 const { obtenerPokemones } = require("../controllers/pokemones");
+const { verifyToken } = require("../middleware/jw-validate");
 const router = express.Router();
 
-router.get("/pokedex", obtenerPokemones);
+router.get("/pokedex", verifyToken , obtenerPokemones);
 
 module.exports = router;
