+++ frontend/src/App.js
@@ -3,10 +3,7 @@ import LoginComponent from "./components/Section/LoginComponent/LoginComponent"
 import Pokedex from "./components/main/PokedexUI/Pokemon";
 import Card from "./components/Section/CardComponent/card-pokemon";
 import "./App.css";
-import POKE_ARRAY from "./data";
 import React, { useState } from "react";
-import ReactDOM from "react-dom/client";
-import auth from "./components/api/ApiAuth";
 import { useEffect } from "react";
 import {
   BrowserRouter,
@@ -20,26 +17,11 @@ import pokemons from "./components/api/ApiPokemones";
 
 function App() {
   
-  const [showPokemons, setShowPokemons] = useState([]);
-
-  async function getPokemons() {
-
-    const data = await pokemons.getPokemones();
-    console.log(data)
-    setShowPokemons(data);
-  }
-  
-
-  useEffect(() => {
-    getPokemons();
-  },[])
-
-
   return (
     <BrowserRouter>
       <Routes>
         <Route path="/register" element={<RegisterComponent />}></Route>
-        <Route path="/pokedex" element={<Pokedex POKE_ARRAY={showPokemons} />}></Route>
+        <Route path="/pokedex" element={<Pokedex/>}></Route>
         <Route path=":nombre" element={<Card />}></Route>
         <Route path="/login" element={<LoginComponent />}></Route>
         <Route path="*" element={<Navigate to='/register' replace />}></Route>
