+++ frontend/src/components/api/Api.js
@@ -4,7 +4,7 @@ const api = axios.create({
     baseURL: "http://localhost:4000",
     headers: {
         "Content-Type": "application/json",
-        "token": localStorage.getItem("token")
+        "auth-token": localStorage.getItem("token")
     }
 })
 export default api;
