+++ frontend/src/components/main/PokedexUI/pokemon.css
@@ -5,8 +5,6 @@
   grid-gap: 6px;
   justify-content: center;
   align-content: center;
-  height: 60vh;
-  margin-top: 2vh;
 }
 
 #bodyUI {
